binjip02
I added some modifications to binjip! I have tried to simplify the basis. The idea is to cover the socket with a playwood structure .Sulla its surface there will be an opening that will allow you to insert the plugs.
I uploaded two versions, one without and one with the antenna!